Move all form-related stuff to a new forms.py file
